fix(bspwm): correct offset in window dimensions
I used slop (https://github.com/naelstrof/slop) to get the dimensions
of the windows, yet when using those numbers in the bspwm rules, the
windows where off by 4 pixels. Either it's a problem with how slop
calculates the width and height of the rectangle, or with how bspwm
rules work.

Not really that big of a deal as I only need to decrement by 4 the
width and height returned by slop when writing rules.

Example: if the slop query for a specific window returned
		1382x83+15+947
	the format being "%wx%h+%x+%y"
	then the dimensions for the bspwm rule should be
		1378x79+15+947