fix(polybar): repair after upgrade to release 3.5
Apparently the new version of polybar brought to light a few issues
existing within my configuration, like the battery module turning green
when fully charged (sounds like a great idea, looks horrendous), and a
few unnecessary lines.

It also introduced the need to specify a separator for the bspwm module
(a simple space).