fix(sxhkd): merge keybindings into a single file
sxhkd needs to be sent the -m option with a -1 for it to work with a
non-qwerty layout, but apparently it only applies to it's first run
instance.

Having 2 files prevented the last one called to work properly. It
doesn't work to send the -m -1 option to each instance. The only way I
found to bind every keymapping to the logical keys (depending on
the current active keyboard layout) was to run sxhkd only once, with a
single config file.