feat($): add "$" command to avoid cli errors
Many times, to represent that some command should be executed as a
normal user on a terminal, a dollar sign is put at the beginning (as if
were the prompt).

Example:
	$ echo "Hello, World!"

As I tend to copy and paste the whole line, an error like `zsh: command
not found: $` is not uncommon.

This scripts solves exactly that, creating the command "$" to execute
"$1" as an executable, with "${@:2}" as it's arguments.